% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect.R
\name{protect_rmd}
\alias{protect_rmd}
\title{Password Protect an R Markdown File}
\usage{
protect_rmd(path, pw, ...)
}
\arguments{
\item{path}{Path to a R Markdown file.}

\item{pw}{Password.}

\item{...}{Additional arguments passed to \code{rmarkdown::render}.}
}
\value{
\code{path}, invisibly.
}
\description{
\code{protect_rmd} provides lightweight password protection for your
HTML format R Markdown reports. Simply point \code{protect_rmd} to the location
of your file and provide a password.
}
