% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect.R
\name{stylize}
\alias{stylize}
\title{Stylize Landing Page}
\usage{
stylize(
  header_text = "Please enter the password",
  incorrect_text = "Incorrect password",
  placeholder = "password",
  button_text = "Continue",
  font_family = "-apple-system",
  background_color = "#f2f2f2",
  font_color = "#2d3737",
  box_color = "#ffffff",
  btn_font_color = "#ffffff",
  btn_bg_color = "#228843",
  btn_hover_color = "#1c6d36",
  alert_font_color = "#a94442",
  alert_bg_color = "#f2dede",
  alert_border_color = "#ebccd1"
)
}
\arguments{
\item{header_text}{Header text.}

\item{incorrect_text}{Incorrect password text.}

\item{placeholder}{Password input placeholder.}

\item{button_text}{Button text.}

\item{font_family}{Font family.}

\item{background_color}{Background color.}

\item{font_color}{Font color.}

\item{box_color}{Box color.}

\item{btn_font_color}{Button font color.}

\item{btn_bg_color}{Button background color.}

\item{btn_hover_color}{Button hover color.}

\item{alert_font_color}{Alert font color.}

\item{alert_bg_color}{Alert background color.}

\item{alert_border_color}{Alert border color.}
}
\value{
A list.
}
\description{
Use this to style the appearance of the landing page
for the protected Rmd file.
}
\examples{

stylize(font_family = "Times")

}
