% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect.R
\name{protect_rmd}
\alias{protect_rmd}
\title{Password protect an Rmd file}
\usage{
protect_rmd(path, pw, ...)
}
\arguments{
\item{path}{the rmd file you want to protect}

\item{pw}{the password}

\item{...}{additional arguments passed to rmarkdown::render}
}
\description{
Password protect an Rmd file
}
