% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect.R
\name{protect_rmd}
\alias{protect_rmd}
\title{Password protect an Rmd file}
\usage{
protect_rmd(path, pw, ...)
}
\arguments{
\item{path}{the rmd file you want to protect}

\item{pw}{the password}

\item{...}{additional arguments passed to rmarkdown::render}
}
\value{
The key used to encrypt the file as an invisible raw vector.
}
\description{
Password protect an Rmd file
}
\details{
Warning: You are using this at your own risk. Make sure your encryption key is
strong enough. For serious use cases, please also review the code of the functions.
Any feedback is appreciated. This is an early package version.
}
\references{
The package follows the same approach as this GitHub repo \href{https://github.com/scottishstoater/protected-github-pages}.
}
