% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect.R
\name{protect_rmd}
\alias{protect_rmd}
\title{Password Protect an R Markdown File}
\usage{
protect_rmd(input, output = NULL, pw, style = stylize(), ...)
}
\arguments{
\item{input}{Path to the input R Markdown file to be rendered.}

\item{output}{Path to where the output files should be saved.}

\item{pw}{Password.}

\item{style}{Object returned from \code{stylize()}.}

\item{...}{Additional arguments passed to \code{rmarkdown::render()}.}
}
\value{
\code{input}, invisibly.
}
\description{
\code{protect_rmd} provides lightweight password protection for your
HTML format R Markdown reports. Simply point \code{protect_rmd} to the location
of your file and provide a password.
}
